avance

+++ Despacho/api/views.py
@@ -12,20 +12,18 @@ from .serializers import EnvioSerializers, EntregaSerializers
 
 # Create your views here.
 
-class EnvioGet(APIView):
-    serializer_class = EnvioSerializers
-    
-    def get(self,request, nroSeguimiento=0):
+@csrf_exempt
+@api_view(['GET','POST','DELETE'])   
+def envio(request, nroSeguimiento=0):
+    if request.method == 'GET':
         if(nroSeguimiento > 0):
             envio=Envio.objects.filter(nroSeguimiento=nroSeguimiento).values()
             if len(envio) > 0:
-                envios = envio[0]
-                datos = {'envio':envios}
+                envio = envio[0]
+                datos = {'envio':envio}
             else:
                 datos={'message':'nroSeguimiento no encontrado'}
             return JsonResponse(datos)
-                
-            
         else:    
             envio=Envio.objects.all()
             if len(envio)>0:
@@ -33,28 +31,15 @@ class EnvioGet(APIView):
             else:
                 datos={'message':'not found'}
             serializer=EnvioSerializers(envio,many=True)  
-            return Response(serializer.data)   
- 
-      
-class EnvioVistaPost(APIView):  
-    serializer_class = EnvioSerializers
-        
-    def post(self, request, *args, **kwargs):
-        envioCrear = request.data
-        print(envioCrear)
-        envioCreado = Envio.objects.create(nombre=envioCrear['nombre'],rut=envioCrear['rut'],direccion=envioCrear['direccion'],
-                             producto=envioCrear['producto'],
-                             cantidad=envioCrear['cantidad'],precio=envioCrear['precio'])
-        
-        envioCreado.save()
-        
-        serializer = EnvioSerializers(envioCreado)
-        
-        return Response(serializer.data)
-
-class EliminarEnvio(APIView):
-    
-    def delete(self, request, nroSeguimiento):
+            return Response(serializer.data)
+    elif request.method == 'POST':
+        serializers = EnvioSerializers(data=request.data)
+        if serializers.is_valid():
+            serializers.save()
+            return Response(serializers.data, status=status.HTTP_201_CREATED)
+        else:
+            return Response(serializers.errors, status=status.HTTP_400_BAD_REQUEST)
+    elif request.method == 'DELETE':
         eliminarEnvio = Envio.objects.get(nroSeguimiento=nroSeguimiento)
         eliminarEnvio.delete()
         datos = {'message':'eliminado con exito'}
